// public/app/core/components/TransformersUI/setGeometry/models.gen.ts
import { FrameGeometrySource, FrameGeometrySourceMode } from '@grafana/data';

// This file should be generated by cue schema

export enum SetGeometryAction {
  Prepare = 'prepare',
  Calculate = 'calculate',
  Modify = 'modify',
}

export enum CalculateFunction {
  Area = 'area',
  Heading = 'heading',
  Distance = 'distance',
}

export enum ModifyFunction {
  AsLine = 'asLine',
  LineTo = 'lineTo',
}

export interface CalculateOptions {
  what?: CalculateFunction;
  field?: string;
}

export interface ModifyOptions {
  fn: ModifyFunction;
  lineTo?: FrameGeometrySource;
}

/** The main transformer options */
export interface SetGeometryOptions {
  action?: SetGeometryAction;
  source?: FrameGeometrySource;
  calculate?: CalculateOptions;
  modify?: ModifyOptions;
}

export const defaultOptions: SetGeometryOptions = {
  action: SetGeometryAction.Prepare,
  source: {
    mode: FrameGeometrySourceMode.Auto,
  },
};